+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "classnames": "^2.5.1",
     "firebase": "^10.13.1",
+    "firebase-functions": "^6.0.1",
     "formik": "^2.4.6",
     "next": "^14.2.6",
     "react": "^18",
