+++ src/components/Header/Header.tsx
@@ -4,6 +4,8 @@ import style from "./Header.module.css";
 import { Button } from "../Controls";
 import Link from "next/link";
 import { useAuthStore } from "@/store/auth";
+import Image from "next/image";
+import img from "../../../public/images/avatar.png"
 
 interface HeaderProps {}
 
@@ -27,8 +29,8 @@ export const Header: FC<HeaderProps> = () => {
           </div>
           {user ? (
             <div className={style.header_avatar_wrapper}>
-              <img
-                src={user.photoURL || "./public/images/avatar"} // Заміна на аватар користувача
+              <Image
+                src={img} // Заміна на аватар користувача
                 alt="User Avatar"
                 className={style.user_avatar}
               />
@@ -36,7 +38,7 @@ export const Header: FC<HeaderProps> = () => {
           ) : (
             <>
               <div className={style.header_btn_wrapper}>
-                <Button href="/#signUpForm" className={style.button}>
+                <Button href="/signup" className={style.button}>
                   Sign up
                 </Button>
               </div>
