+++ src/components/JoinUsForm/JoinUsForm.module.css
@@ -311,17 +311,3 @@
   border-top: none;
   border-bottom: none;
 }
-
-.success_title {
-  font-size: 40px;
-  font-weight: 400;
-  line-height: 40px;
-  text-align: center;
-}
-
-.success_text {
-  font-size: 25px;
-  font-weight: 400;
-  line-height: 40px;
-  text-align: center;
-}
