+++ src/components/SignUpForm/SignUpForm.tsx
@@ -1,31 +1,20 @@
 /* eslint-disable @next/next/no-img-element */
-import React, { useState, useEffect, useRef, ChangeEvent, FC } from "react";
-import classNames from "classnames";
-import style from "./signUpForm.module.css";
-import { Button } from "../Controls/Button";
-import { RadioButton } from "../Radio/Radio";
-import { Input } from "../Input/Input";
-import { Preloader } from "../Preloader/Preloader";
-import { InputMasked } from "../Input/InputMasked";
+import React, { FC } from "react";
 import { useFormik, FormikHelpers } from "formik";
 import * as Yup from "yup";
-import useImageValidation from "../../hooks/useImageValidation";
-import useApiRequest from "../../hooks/useRequest";
 import {
   createUserWithEmailAndPassword,
   UserCredential,
   AuthError,
-  getAuth, // Додаємо getAuth з Firebase
+  getAuth,
 } from "firebase/auth";
+import Image from "next/image";
+import img from "../../../public/images/Success.png"
 import { useAuthStore } from "../../store/auth/useAuthStore";
-import {
-  confirmPasswordValidationRule,
-  emailValidationRule,
-  nameValidationRule,
-  passwordValidationRule,
-  photoValidationRule,
-  positonValidationRule,
-} from "@/utils/validationRules";
+import { Input } from "../Input/Input";
+import { Button } from "../Controls/Button";
+import style from "./signUpForm.module.css";
+import { emailValidationRule, passwordValidationRule, nameValidationRule, confirmPasswordValidationRule } from "../../utils/validationRules";
 
 // Ініціалізуємо Firebase auth
 const auth = getAuth();
@@ -34,90 +23,31 @@ const auth = getAuth();
 interface FormValues {
   name: string;
   email: string;
-  phone: string;
-  position: string;
   password: string;
   confirmPassword: string;
-  photo: File | null;
 }
 
-// Interface for positon
-interface Position {
-  id: string;
-  name: string;
-}
-
-// main form component
+// Main form component
 export const SignUpForm: FC = () => {
-  const fileInputRef = useRef<HTMLInputElement | null>(null);
-  const [positions, setPositions] = useState<Position[]>([]);
-  const [selectedPosition, setSelectedPosition] = useState<string>("");
-  const [isSuccess, setIsSuccess] = useState<boolean>(false);
-  const { imageError, fileName, handleFileChange } = useImageValidation({});
-
   const { setUser } = useAuthStore();
-
-  // Hook for post request
-  const {
-    makeRequest: makePostRequest,
-    loading: postLoading,
-    error: postError,
-  } = useApiRequest<FormData>(
-    "https://frontend-test-assignment-api.abz.agency/api/v1/users",
-    "POST"
-  );
-
-  // Hook for get request
-  const {
-    makeRequest: makeGetRequest,
-    loading: getLoading,
-    error: getError,
-  } = useApiRequest<void>(
-    "https://frontend-test-assignment-api.abz.agency/api/v1/positions",
-    "GET"
-  );
-
-  useEffect(() => {
-    const fetchPositions = async () => {
-      try {
-        const data = await makeGetRequest(); // Use get request with hook
-        if (data.positions) {
-          setPositions(data.positions);
-          if (data.positions.length > 0) {
-            const defaultPosition = data.positions[0].id;
-            setSelectedPosition(String(defaultPosition));
-            setFieldValue("position", defaultPosition);
-          }
-        }
-      } catch (error) {
-        console.error("Ошибка при получении позиций:", getError);
-      }
-    };
-
-    fetchPositions();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  const [isSuccess, setIsSuccess] = React.useState<boolean>(false);
 
   const initialValues: FormValues = {
     name: "",
     email: "",
-    phone: "",
-    position: selectedPosition,
     password: "",
     confirmPassword: "",
-    photo: null,
   };
 
-  const validationSchema = Yup.object({
-    name: nameValidationRule(),
-    email: emailValidationRule(),
-    phone: photoValidationRule(),
-    position: positonValidationRule(),
-    password: passwordValidationRule(),
-    confirmPassword: confirmPasswordValidationRule(),
-    photo: photoValidationRule(),
-  });
+  // Validation schema using Yup
+ const validationSchema = Yup.object({
+   name: nameValidationRule(),
+   email: emailValidationRule(),
+   password: passwordValidationRule(),
+   confirmPassword: confirmPasswordValidationRule(),
+ });
 
+  // Formik hook
   const {
     values,
     errors,
@@ -125,7 +55,6 @@ export const SignUpForm: FC = () => {
     handleChange,
     handleBlur,
     handleSubmit,
-    setFieldValue,
     isValid,
     dirty,
   } = useFormik({
@@ -135,68 +64,36 @@ export const SignUpForm: FC = () => {
       values: FormValues,
       { setFieldError }: FormikHelpers<FormValues>
     ) => {
-      const formData = new FormData();
-      formData.append("name", values.name);
-      formData.append("email", values.email);
-      formData.append("phone", values.phone);
-      formData.append("position_id", selectedPosition);
-      formData.append("photo", values.photo as Blob);
-
       try {
         const userCredential: UserCredential =
           await createUserWithEmailAndPassword(
-            auth, // Замінили Auth на правильно ініціалізований об'єкт auth
+            auth,
             values.email,
-            values.password // Using password for authentication
+            values.password
           );
         const user = userCredential.user;
-        setUser(user); // save user in Zustand
-        await makePostRequest(formData); // Call post request with hook
+        setUser(user); // Зберігаємо користувача в Zustand
         setIsSuccess(true);
       } catch (error) {
         const authError = error as AuthError;
-        setFieldError("email", authError.message || "Ошибка при регистрации");
+        setFieldError(
+          "email",
+          authError.message || "Error during registration"
+        );
       }
     },
   });
 
-  const handleUpload = () => {
-    if (fileInputRef.current) {
-      fileInputRef.current.click();
-    }
-  };
-
-  const handleFileInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (file) {
-      handleFileChange(file, setFieldValue);
-    }
-  };
-
-  const getShortFileName = (name: string) => {
-    const maxLength = 20;
-    if (name.length <= maxLength) {
-      return name;
-    }
-    return `${name.substring(0, 7)}...${name.substring(name.length - 10)}`;
-  };
-
   return (
     <>
-      {postLoading || getLoading ? (
-        <Preloader />
-      ) : isSuccess ? (
+      {isSuccess ? (
         <div className={style.success_screen}>
-          <h4 className={style.success_title}>User successfully registered</h4>
-          <img
-            src={"./images/Success.png"}
-            alt="Success"
-            className={style.success_img}
-          />
+                  <h4 className={style.success_title}>Thank you for registration</h4>
+                  <Image alt="" src={img}/>
         </div>
       ) : (
         <form className={style.form} id="signUpForm" onSubmit={handleSubmit}>
-          <h3 className={style.form_title}>Working with POST request</h3>
+          <h3 className={style.form_title}>Sign Up</h3>
           <div className={style.form_content_wrapper}>
             <Input
               id="name"
@@ -218,16 +115,6 @@ export const SignUpForm: FC = () => {
               name="email"
               errorMessage={touched.email && errors.email}
             />
-            <InputMasked
-              id="phone"
-              value={values.phone}
-              type="tel"
-              placeholder="Phone"
-              onChange={handleChange}
-              onBlur={handleBlur}
-              name="phone"
-              errorMessage={touched.phone && errors.phone}
-            />
             <Input
               id="password"
               value={values.password}
@@ -248,59 +135,10 @@ export const SignUpForm: FC = () => {
               name="confirmPassword"
               errorMessage={touched.confirmPassword && errors.confirmPassword}
             />
-            <div className={style.radio_wrapper}>
-              <h2 className={style.radio_buttons__title}>
-                Select your position
-              </h2>
-              {positions.map((position) => (
-                <RadioButton
-                  key={position.id}
-                  position={position}
-                  selectedPosition={selectedPosition}
-                  onChange={(value) => {
-                    setSelectedPosition(value);
-                    setFieldValue("position", value);
-                  }}
-                />
-              ))}
-            </div>
-            <div
-              className={classNames(style.upload_container, {
-                [style.error_border]: imageError,
-              })}
-            >
-              <button
-                type="button"
-                onClick={handleUpload}
-                className={classNames(style.upload_button, {
-                  [style.error_btn_input]: imageError,
-                })}
-              >
-                Upload
-              </button>
-              <input
-                ref={fileInputRef}
-                className={style.img_input}
-                type="file"
-                name="photo"
-                onChange={handleFileInputChange}
-                onBlur={handleBlur}
-                style={{ display: "none" }}
-              />
-              <div className={style.image_preview_container}>
-                {fileName ? getShortFileName(fileName) : "Upload your photo"}
-              </div>
-              {imageError && (
-                <div className={style.error_message}>{imageError}</div>
-              )}
-            </div>
-            {(postError || getError) && (
-              <div className={style.error_message}>{postError || getError}</div>
-            )}
             <div className={style.form_btn_wrapper}>
               <Button
                 type="submit"
-                disabled={!isValid || !dirty || postLoading}
+                disabled={!isValid || !dirty}
                 className={style.btn_submit}
               >
                 Sign up
