+++ src/pages/signup.tsx
@@ -1,6 +1,10 @@
-import { UploadImageForm } from "@/components/SignUpForm/signUp-form";
+import { SignUpForm } from "../components/SignUpForm/SignUpForm";
 import style from "./login.module.scss";
 
 export default function SignUp() {
-  return <div className={style.login_wrapper}>{/* <UploadImageForm /> */}</div>;
+  return (
+    <div className={style.login_wrapper}>
+      <SignUpForm />
+    </div>
+  );
 }
