+++ src/store/auth/useAuthStore.ts
@@ -1,10 +1,13 @@
 import { create } from "zustand";
-import { getAuth, signInWithEmailAndPassword, createUserWithEmailAndPassword } from "firebase/auth";
-
-const auth = getAuth(); // Імпортуємо Firebase auth
+import {
+  signInWithEmailAndPassword,
+  createUserWithEmailAndPassword,
+  User,
+} from "firebase/auth";
+import { auth } from "./firebase.config";
 
 interface IAuthStore {
-  user: any;
+  user: User | null;
   setUser: (user: any) => void;
   login: ({ email, password }: IAuthCredentials) => Promise<void>;
   signUp: ({ email, password }: IAuthCredentials) => Promise<void>;
@@ -17,16 +20,20 @@ interface IAuthCredentials {
 
 export const useAuthStore = create<IAuthStore>((set) => ({
   user: null,
-  
+
   setUser: (user: any) => set({ user }),
 
   login: async ({ email, password }: IAuthCredentials) => {
     const { user } = await signInWithEmailAndPassword(auth, email, password); // Використовуємо об'єкт auth
     set({ user });
   },
-  
+
   signUp: async ({ email, password }: IAuthCredentials) => {
-    const { user } = await createUserWithEmailAndPassword(auth, email, password); // Використовуємо об'єкт auth
+    const { user } = await createUserWithEmailAndPassword(
+      auth,
+      email,
+      password
+    ); // Використовуємо об'єкт auth
     set({ user });
   },
 }));
